requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
deepl==1.5.0
discord-ext-ui==3.1.3
flake8==4.0.1
frozenlist==1.3.0
idna==3.3
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
newdispanderfixed==0.1.5
pathspec==0.9.0
pipdeptree==2.2.1
platformdirs==2.5.1
git+https://github.com/Pycord-Development/pycord@f4b0d07388d2a4bd6ca98b3b10a9cab7d64ca34e
pycodestyle==2.8.0
pyflakes==2.4.0
requests==2.27.1
tomli==2.0.1
urllib3==1.26.8
yarl==1.7.2
